Fix mission and sos-requests to fit new DB and using TransactionalEventListener

+++ spring-backend/src/main/java/com/drc/aidbridge/modules/sos/SosRequestCreatedEvent.java
@@ -5,12 +5,12 @@ import lombok.Builder;
 import lombok.Getter;
 
 import java.util.UUID;
-
+import java.math.BigDecimal;
 @Getter
 @Builder
 @AllArgsConstructor
 public class SosRequestCreatedEvent {
     private final UUID sosRequestId;
-    private final Double lat;
-    private final Double lng;
+    private final BigDecimal lat;
+    private final BigDecimal lng;
 }
